requirements.txt
bcrypt==3.2.0
certifi>=2022.12.07
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
mysqlclient==2.1.0
psutil==5.9.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
python-http-client==3.3.7
python-nmap==0.7.1
requests==2.27.1
sendgrid==6.9.7
six==1.16.0
SQLAlchemy==1.4.32
starkbank-ecdsa==2.0.3
urllib3==1.26.9
Werkzeug>=2.2.3
